fix: refactored and made new components for sort by and filter by button
