+++ src/community/common/assets/languages/english/common.json
@@ -31,7 +31,7 @@
       "noTeamTxt": "No teams assigned "
     },
     "dateRangePicker": {
-      "pickerLabel": "Select Date"
+      "label": "Date: "
     },
     "calendarDataRangePicker": {
       "label": "Select date(s)",
